requirements.txt
torch==2.6.0
torchvision>=0.15.0
typer==0.15.2
fastapi==0.115.12
uvicorn==0.34.2
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=9.5.0
pyyaml>=6.0
tqdm>=4.65.0
pathlib>=1.0.1
click>=8.0.0
rich>=13.0.0
scipy>=1.10.0
scikit-image>=0.20.0
open3d>=0.17.0
